LayoutTests/editing/spelling/spelling-should-select-multiple-words-expected.txt

Spell check should select multi-word misspellings on context click. To test manually, enable 'Ask Google for Suggestions' in Chrome, type 'It should be upper case.', and context-click on the word 'upper'. The test succeeds when 'upper case' was selected after context click.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spellingMarkerRange.toString() is "upper case"
PASS window.getSelection().toString() is ""
PASS spellingMarkerRange.toString() is "upper case"
PASS window.getSelection().toString() is "upper case"
PASS contextMenuElements[contextMenuElements.length - 1] is "uppercase"
PASS successfullyParsed is true

TEST COMPLETE
It should be upper case.
